+++ api/models/track_properties.py
@@ -29,12 +29,6 @@ class Provenance(BaseModel):
     release_date: Optional[date] 
     
 
-class FILERAccession(Provenance):
-    data_source_version: Optional[str]
-    download_date: Optional[date] 
-    project: Optional[str] # FIXME: make this == collections: List[str]?
-    
-    
 class DSSAccession(Provenance):
     data_source: str = "NIAGADS DSS"
     accession: str
